+++ packages/backend-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @windrun-huaiin/backend-core
 
+## 14.3.0
+
+### Minor Changes
+
+- feat(dependency): update dependencies
+  - optimize: `diaomao-update` cli command
+  - security: user mock or test now fully split with real code
+  - fix: fingerprint panel test now won't override origin fpId, with new local storage
+
+### Patch Changes
+
+- Updated dependencies
+  - @windrun-huaiin/third-ui@14.2.0
+  - @windrun-huaiin/lib@14.0.1
+
 ## 14.2.0
 
 ### Minor Changes